requirements.txt
flask
openai
python-dotenv
flask_session
PyPDF2